rename files (remove Tr)
