+++ src/frontend/components/Commands/Effects/AddStreamTagsEffect.vue
