+++ src/frontend/components/Commands/Effects/ChatEffect.vue
