+++ src/frontend/components/Commands/Effects/DictLookupEffect.vue
