+++ src/frontend/components/Commands/Effects/EmotesEffect.vue
