+++ src/frontend/components/Commands/Effects/MadochanEffect.vue
