+++ src/frontend/components/Commands/Effects/MediaEffect.vue
