+++ src/frontend/components/Commands/Effects/RemoveStreamTagsEffect.vue
