+++ src/frontend/components/Commands/Effects/SetChannelGameIdEffect.vue
