+++ src/frontend/components/Commands/Effects/SetChannelTitleEffect.vue
