+++ src/frontend/components/Commands/Effects/VariableChangeEffect.vue
