+++ src/frontend/components/Commands/EffectsEditor.vue
@@ -7,55 +7,55 @@
       >
         <td><code>{{ item.type }}</code></td>
         <td>
-          <TrVariableChangeEffect
+          <VariableChangeEffect
             v-if="item.type === CommandEffectType.VARIABLE_CHANGE"
             v-model="val[idx]"
             :item-variables="itemVariables"
             :global-variables="globalVariables"
           />
-          <TrChatEffect
+          <ChatEffect
             v-if="item.type === CommandEffectType.CHAT"
             v-model="val[idx]"
           />
-          <TrDictLookupEffect
+          <DictLookupEffect
             v-if="item.type === CommandEffectType.DICT_LOOKUP"
             v-model="val[idx]"
           />
-          <TrEmotesEffect
+          <EmotesEffect
             v-if="item.type === CommandEffectType.EMOTES"
             v-model="val[idx]"
           />
-          <TrMediaEffect
+          <MediaEffect
             v-if="item.type === CommandEffectType.MEDIA"
             v-model="val[idx]"
             :base-volume="baseVolume"
             :widget-url="widgetUrl"
           />
-          <TrMadochanEffect
+          <MadochanEffect
             v-if="item.type === CommandEffectType.MADOCHAN"
             v-model="val[idx]"
           />
-          <TrSetChannelTitleEffect
+          <SetChannelTitleEffect
             v-if="item.type === CommandEffectType.SET_CHANNEL_TITLE"
             v-model="val[idx]"
           />
-          <TrSetChannelGameIdEffect
+          <SetChannelGameIdEffect
             v-if="item.type === CommandEffectType.SET_CHANNEL_GAME_ID"
             v-model="val[idx]"
           />
-          <TrAddStreamTagsEffect
+          <AddStreamTagsEffect
             v-if="item.type === CommandEffectType.ADD_STREAM_TAGS"
             v-model="val[idx]"
           />
-          <TrRemoveStreamTagsEffect
+          <RemoveStreamTagsEffect
             v-if="item.type === CommandEffectType.REMOVE_STREAM_TAGS"
             v-model="val[idx]"
           />
-          <TrChattersEffect
+          <ChattersEffect
             v-if="item.type === CommandEffectType.CHATTERS"
             v-model="val[idx]"
           />
-          <TrCountdownEffect
+          <CountdownEffect
             v-if="item.type === CommandEffectType.COUNTDOWN"
             v-model="val[idx]"
             :base-volume="baseVolume"
@@ -78,18 +78,18 @@
 <script setup lang="ts">
 import { ref, watch } from 'vue';
 import { CommandEffect, CommandEffectType, CommandVariable, GlobalVariable } from '../../../types';
-import TrVariableChangeEffect from './Effects/TrVariableChangeEffect.vue';
-import TrChatEffect from './Effects/TrChatEffect.vue';
-import TrDictLookupEffect from './Effects/TrDictLookupEffect.vue';
-import TrEmotesEffect from './Effects/TrEmotesEffect.vue';
-import TrMediaEffect from './Effects/TrMediaEffect.vue';
-import TrMadochanEffect from './Effects/TrMadochanEffect.vue';
-import TrSetChannelTitleEffect from './Effects/TrSetChannelTitleEffect.vue';
-import TrSetChannelGameIdEffect from './Effects/TrSetChannelGameIdEffect.vue';
-import TrAddStreamTagsEffect from './Effects/TrAddStreamTagsEffect.vue';
-import TrRemoveStreamTagsEffect from './Effects/TrRemoveStreamTagsEffect.vue';
-import TrChattersEffect from './Effects/TrChattersEffect.vue';
-import TrCountdownEffect from './Effects/TrCountdownEffect.vue';
+import VariableChangeEffect from './Effects/VariableChangeEffect.vue';
+import ChatEffect from './Effects/ChatEffect.vue';
+import DictLookupEffect from './Effects/DictLookupEffect.vue';
+import EmotesEffect from './Effects/EmotesEffect.vue';
+import MediaEffect from './Effects/MediaEffect.vue';
+import MadochanEffect from './Effects/MadochanEffect.vue';
+import SetChannelTitleEffect from './Effects/SetChannelTitleEffect.vue';
+import SetChannelGameIdEffect from './Effects/SetChannelGameIdEffect.vue';
+import AddStreamTagsEffect from './Effects/AddStreamTagsEffect.vue';
+import RemoveStreamTagsEffect from './Effects/RemoveStreamTagsEffect.vue';
+import ChattersEffect from './Effects/ChattersEffect.vue';
+import CountdownEffect from './Effects/CountdownEffect.vue';
 
 export interface AutocompletableVariable {
   var: CommandVariable | GlobalVariable;
